cv_service/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
pika==1.3.2
PyMuPDF==1.24.10
requests==2.32.3
